Adding basic token validation to all actions

+++ client/src/Actions/PaymentMethodActions.js
@@ -2,10 +2,23 @@ import axios from 'axios';
 import { URL } from './index';
 import * as types from './actionTypes';
 
+import { tokenIsValid } from './../Helpers/tokenIsValid'
+
 export const fetchPaymentMethods = (token) => dispatch => {
   dispatch({
     type: types.FETCHING_PAYMENT_METHODS_START
   })
+
+  // Preliminary token validation
+  if (!tokenIsValid(token)) {
+    return dispatch({ 
+      type: types.FETCHING_PAYMENT_METHODS_FAILURE, 
+      payload: {
+        error: "The provided token is invalid, I'm afraid! Make sure it's a string of the appropriate length"
+      } 
+    });
+  }
+
   return axios.get(`${URL}/api/me/payment-methods`, { 
     headers: {
       'Content-Type': 'application/json',
@@ -35,6 +48,17 @@ export const fetchIndividualPaymentMethod = (payment_method_id, token) => dispat
   dispatch({
     type: types.FETCHING_PAYMENT_METHODS_START
   })
+
+  // Preliminary token validation
+  if (!tokenIsValid(token)) {
+    return dispatch({ 
+      type: types.FETCHING_PAYMENT_METHODS_FAILURE, 
+      payload: {
+        error: "The provided token is invalid, I'm afraid! Make sure it's a string of the appropriate length"
+      } 
+    });
+  }
+
   return axios.get(`${URL}/api/me/payment-methods/${payment_method_id}`, { 
     headers: {
       'Content-Type': 'application/json',
@@ -72,6 +96,16 @@ export const addPaymentMethod = (new_payment_menthod, token) => dispatch => {
     type: types.ADDING_PAYMENT_METHOD_START
   })
 
+  // Preliminary token validation
+  if (!tokenIsValid(token)) {
+    return dispatch({ 
+      type: types.ADDING_PAYMENT_METHOD_FAILURE, 
+      payload: {
+        error: "The provided token is invalid, I'm afraid! Make sure it's a string of the appropriate length"
+      } 
+    });
+  }
+
   return axios.post(`${URL}/api/me/payment-methods`, {
     headers: {
       'Content-Type': 'application/json',
@@ -102,6 +136,16 @@ export const removePaymentMethod = (payment_method_id, id, token) => dispatch =>
   dispatch({
     type: types.REMOVING_PAYMENT_METHOD_START
   })
+
+  // Preliminary token validation
+  if (!tokenIsValid(token)) {
+    return dispatch({ 
+      type: types.REMOVING_PAYMENT_METHOD_FAILURE, 
+      payload: {
+        error: "The provided token is invalid, I'm afraid! Make sure it's a string of the appropriate length"
+      } 
+    });
+  }
 
   return axios.delete(`${URL}/api/payment-methods/${payment_method_id}`, { 
     headers: {
